requirements.txt
asgiref==3.5.2
Babel==2.10.3
Django==4.0.6
django-money==3.0.0
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
psycopg2==2.9.3
py-moneyed==2.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
